+++ pymongo/operations.py
@@ -243,6 +243,8 @@ class UpdateOne(_UpdateOp):
           - `array_filters` (optional): A list of filters specifying which
             array elements an update should apply. Requires MongoDB 3.6+.
 
+        .. versionchanged:: 3.9
+           Added the ability to accept a pipeline as the `update`.
         .. versionchanged:: 3.6
            Added the `array_filters` option.
         .. versionchanged:: 3.5
@@ -280,6 +282,8 @@ class UpdateMany(_UpdateOp):
           - `array_filters` (optional): A list of filters specifying which
             array elements an update should apply. Requires MongoDB 3.6+.
 
+        .. versionchanged:: 3.9
+           Added the ability to accept a pipeline as the `update`.
         .. versionchanged:: 3.6
            Added the `array_filters` option.
         .. versionchanged:: 3.5